tweak debug plot

+++ pdf2docx/converter.py
@@ -16,7 +16,7 @@ class Converter:
         with python-docx.
     '''
 
-    def __init__(self, pdf_file:str, docx_file:str=None, debug:bool=False) -> None:
+    def __init__(self, pdf_file:str, docx_file:str=None, debug:bool=False):
         ''' Initialize fitz object with given pdf file path; initialize docx object.
 
             If debug=True, illustration pdf will be created during parsing the raw pdf layout.
@@ -24,8 +24,7 @@ class Converter:
         # pdf/docx filename
         self.filename_pdf = pdf_file
         self.filename_docx = docx_file if docx_file else pdf_file.replace('.pdf', '.docx')
-        if os.path.exists(self.filename_docx):
-            os.remove(self.filename_docx)
+        if os.path.exists(self.filename_docx): os.remove(self.filename_docx)
 
         # fitz object to read pdf
         self._doc_pdf = fitz.open(pdf_file)
@@ -38,13 +37,13 @@ class Converter:
 
         # fitz object in debug mode: plot page layout
         # file path for this debug pdf: demo.pdf -> debug_demo.pdf
-        self.debug_mode = debug
         path, filename = os.path.split(pdf_file)
-        self.filename_debug = os.path.join(path, f'debug_{filename}')
-        self._doc_debug = fitz.open() if debug else None
-
-        # to serialize layout for debug purpose
-        self._filename_debug_layout = os.path.join(path, 'layout.json')
+        self._debug_kwargs = {
+            'debug'   : debug,
+            'doc'     : fitz.open() if debug else None,
+            'filename': os.path.join(path, f'debug_{filename}'),
+            'layout'  : os.path.join(path, 'layout.json')
+        }
 
 
     def __getitem__(self, index):
@@ -58,88 +57,45 @@ class Converter:
         else:
             return self._doc_pdf[index]
 
+
     def __len__(self):
         return len(self._doc_pdf)
 
-    @property
-    def layout(self):
-        return self._layout
 
     @property
-    def doc_pdf(self):
-        return self._doc_pdf
+    def layout(self): return self._layout
 
     @property
-    def doc_docx(self):
-        return self._doc_docx
+    def doc_pdf(self): return self._doc_pdf
 
     @property
-    def _debug_kwargs(self):
-        return {
-                'debug': self.debug_mode,
-                'doc': self._doc_debug,
-                'filename': self.filename_debug
-            }
-
-    def init(self, page:fitz.Page) -> Layout:
-        '''Initialize layout object.'''
-        # Layout object based on raw dict
-        # NOTE: all these coordinates are relative to un-rotated page
-        # https://pymupdf.readthedocs.io/en/latest/page.html#modifying-pages
-        raw_layout = page.getText('rawdict')
-
-        # though 'width', 'height' are contained in `raw_dict`, they are based on un-rotated page.
-        # so, update page width/height to right direction in case page is rotated
-        *_, w, h = page.rect # always reflecting page rotation
-        raw_layout.update({
-            'width' : w,
-            'height': h
-        })
-        self._layout = Layout(raw_layout, page.rotationMatrix)
-        
-        # get rectangle shapes from page source
-        self._layout.rects.from_stream(self.doc_pdf, page)
-        
-        # get annotations(comment shapes) from PDF page, e.g. 
-        # highlight, underline and strike-through-line        
-        self._layout.rects.from_annotations(page)
-
-        # plot raw layout
-        if self.debug_mode:
-            # new section for current pdf page
-            new_page_section(self._doc_debug, self._layout.width, self._layout.height, f'Page {page.number}')
-
-            # initial layout
-            self._layout.plot(self._doc_debug, 'Original Text Blocks', key=PlotControl.LAYOUT)
-            self._layout.plot(self._doc_debug, 'Original Rectangle Shapes', key=PlotControl.SHAPE)
-
-        return self._layout
+    def doc_docx(self): return self._doc_docx
 
 
-    def parse(self, page:fitz.Page):
-        '''Parse page layout.'''
+    def make_page(self, page:fitz.Page):
+        '''Parse pdf page and create docx page.'''
         # parse page: text/table/layout format
-        self.init(page).parse(**self._debug_kwargs)
+        self._init(page).parse(**self._debug_kwargs).make_page(self._doc_docx)
 
-        # debug:
-        # - save layout plotting as pdf file
-        # - write layout information
-        if self.debug_mode:
-            self._doc_debug.save(self.filename_debug)
-            self._layout.serialize(self._filename_debug_layout)
+        # save files
+        self.save_docx()
+        self._debug_save()
 
         return self
 
 
-    def make_page(self):
-        '''Create docx page based on parsed layout.'''
-        self._layout.make_page(self._doc_docx)
-        self.save_docx()
-
-
     def extract_tables(self, page:fitz.Page):
         '''Extract table contents.'''
-        return self.init(page).extract_tables()
+        return self._init(page).extract_tables()
+
+
+    def make_docx(self, page_indexes:list):
+        '''Parse and create a list of pages.
+            ---
+            Args:
+            - page_indexes: list[int], page indexes to parse
+        '''
+        pass
 
 
     def save_docx(self):
@@ -150,5 +106,69 @@ class Converter:
     def close(self):
         '''Close pdf files.'''
         self._doc_pdf.close()
-        if self.debug_mode:
-            self._doc_debug.close()
+        if self._debug_kwargs['debug']: self._debug_kwargs['doc'].close()
+
+
+    def _init(self, page:fitz.Page):
+        '''Initialize layout object.'''
+        # Layout object based on raw dict
+        # NOTE: all these coordinates are relative to un-rotated page
+        # https://pymupdf.readthedocs.io/en/latest/page.html#modifying-pages
+        raw_layout = page.getText('rawdict')
+
+        # though 'width', 'height' are contained in `raw_dict`, they are based on un-rotated page.
+        # so, update page width/height to right direction in case page is rotated
+        *_, w, h = page.rect # always reflecting page rotation
+        raw_layout.update({ 'width' : w, 'height': h })
+        self._layout = Layout(raw_layout, page.rotationMatrix)
+        
+        # get rectangle shapes from page source
+        self._layout.rects.from_stream(self.doc_pdf, page)
+        
+        # get annotations(comment shapes) from PDF page, e.g. 
+        # highlight, underline and strike-through-line        
+        self._layout.rects.from_annotations(page)
+
+        # plot raw layout
+        self._plot_initial_layout(page.number)
+
+        return self._layout
+
+
+    def _make_docx(self, page_indexes:list):
+        '''Parse and create a list of pages.
+            ---
+            Args:
+            - page_indexes: list[int], page indexes to parse
+        '''
+        for i in page_indexes:
+            pass
+
+    def _make_docx_multi_processing(self):
+        pass
+
+
+    def _plot_initial_layout(self, page_number:int):
+        '''Plot initial layout.'''
+        if not self._debug_kwargs['debug']: return
+
+        doc = self._debug_kwargs['doc']
+
+        # new section for current pdf page
+        new_page_section(doc, self._layout.width, self._layout.height, f'Page {page_number}')
+
+        # initial layout
+        self._layout.plot(doc, 'Original Text Blocks', key=PlotControl.LAYOUT)
+        self._layout.plot(doc, 'Original Rectangle Shapes', key=PlotControl.SHAPE)
+
+
+    def _debug_save(self):
+        if not self._debug_kwargs['debug']: return
+        # save layout plotting as pdf file
+        self._debug_kwargs['doc'].save(self._debug_kwargs['filename'])
+        # write layout information
+        self._layout.serialize(self._debug_kwargs['layout'])
+
+
+
+
